refactor(menu): move pure accelerator utilities out of menu.ts (#1906)
formatAccelerator, collectAcceleratorsByMenu, and walkInto were pure and
Electron-free — the doc comment already said so — but lived in menu.ts,
forcing their test to mock the entire Electron + project-state surface
menu.ts pulls in just to reach them.

Moves the three to src/main/menu/accelerators.ts (only a type-only `import
type { MenuItemConstructorOptions } from 'electron'`, erased at build time,
so the module has no runtime Electron dependency).

Splits the test file accordingly:
- tests/main/menu-accelerators.test.ts now imports directly from
  menu/accelerators.ts and drops every mock — collectAcceleratorsByMenu and
  formatAccelerator's own unit tests need none.
- tests/main/menu-shortcuts.test.ts (new) keeps the Electron + project-state
  mocks for the two tests that are genuinely impure: the production-menu
  collision check and getMenuShortcuts both call the real rebuildMenu(),
  which still lives in menu.ts and still needs the full mock surface.

1072 -> 1020 lines. BUDGETS entry lowered in the same diff.
