+++ src/main/menu.ts
@@ -1,5 +1,6 @@
 import { Menu, shell, dialog, BrowserWindow } from 'electron';
 import path from 'node:path';
+import { formatAccelerator, collectAcceleratorsByMenu } from './menu/accelerators';
 import { Channels } from '../shared/channels';
 import { broadcast } from './ipc/broadcast';
 import type { EventMap } from '../shared/ipc-contract';
@@ -1016,57 +1017,4 @@ export function getMenuShortcuts(): ShortcutGroup[] {
     groups.push({ menu, items });
   }
   return groups;
-}
-
-/** Render an Electron accelerator string for display on the current platform
- *  (⌘⇧S on macOS, Ctrl+Shift+S elsewhere). */
-export function formatAccelerator(accelerator: string, platform: NodeJS.Platform = process.platform): string {
-  const isMac = platform === 'darwin';
-  const mac: Record<string, string> = {
-    CmdOrCtrl: '⌘', Cmd: '⌘', Command: '⌘', Ctrl: '⌃', Control: '⌃',
-    Alt: '⌥', Option: '⌥', Shift: '⇧', Super: '⌘', Plus: '+', Minus: '−',
-  };
-  const other: Record<string, string> = {
-    CmdOrCtrl: 'Ctrl', Cmd: 'Ctrl', Command: 'Ctrl', Ctrl: 'Ctrl', Control: 'Ctrl',
-    Alt: 'Alt', Option: 'Alt', Shift: 'Shift', Super: 'Win', Plus: '+', Minus: '−',
-  };
-  const map = isMac ? mac : other;
-  const tokens = accelerator.split('+').map((t) => map[t] ?? t);
-  return isMac ? tokens.join('') : tokens.join('+');
-}
-
-/**
- * Walk a menu template tree and collect every accelerator under each
- * top-level menu. Returns a Map keyed by top-level menu label. Pure;
- * no Electron runtime dependency. Used by the accelerator-collision
- * test (#398).
- */
-export function collectAcceleratorsByMenu(
-  template: Electron.MenuItemConstructorOptions[],
-): Map<string, Array<{ accelerator: string; path: string[] }>> {
-  const out = new Map<string, Array<{ accelerator: string; path: string[] }>>();
-  for (const top of template) {
-    const topLabel = String(top.label ?? top.role ?? '(unnamed)');
-    const found: Array<{ accelerator: string; path: string[] }> = [];
-    walkInto(top, [topLabel], found);
-    if (found.length > 0) out.set(topLabel, found);
-  }
-  return out;
-}
-
-function walkInto(
-  item: Electron.MenuItemConstructorOptions,
-  path: string[],
-  out: Array<{ accelerator: string; path: string[] }>,
-): void {
-  if (typeof item.accelerator === 'string') {
-    out.push({ accelerator: item.accelerator, path });
-  }
-  const sub = item.submenu;
-  if (Array.isArray(sub)) {
-    for (const child of sub) {
-      const childLabel = String(child.label ?? child.role ?? '(unnamed)');
-      walkInto(child, [...path, childLabel], out);
-    }
-  }
 }
