+++ tests/architecture/file-size-budgets.test.ts
@@ -63,7 +63,7 @@ const BUDGETS: Record<string, number> = {
   'src/renderer/lib/components/Editor.svelte': 829,
   'src/renderer/lib/stores/editor.svelte.ts': 1183,
     'src/renderer/lib/components/right-sidebar/PropertiesPanel.svelte': 1156,
-  'src/main/menu.ts': 1072,
+  'src/main/menu.ts': 1020,
   'src/main/graph/indexers.ts': 723,
   'src/renderer/lib/components/Sidebar.svelte': 998,
   'src/renderer/lib/app/refactor-ops.svelte.ts': 827,
